Log sarama messages to a separate set of files and rotate.
Found that there is a large volume of sarama internal logging going to
syslog. This caused the root disk to fill up on staging.

We might want to look at these logs, they could be important to debug a
problem with the consumer, but we don't need to send them to syslog.
Instead, lumberjack is used to rotate and purge sarama logs based on
file size and time.

The snap is configured to log sarama output to a snap-writable location.